add bcrypt and JWT
