+++ server/package.json
@@ -18,11 +18,15 @@
     "typescript": "^4.1.5"
   },
   "dependencies": {
+    "@types/bcryptjs": "^2.4.2",
     "@types/cors": "^2.8.10",
     "@types/express": "^4.17.11",
+    "@types/jsonwebtoken": "^8.5.1",
     "@types/nodemailer": "^6.4.1",
+    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "jsonwebtoken": "^8.5.1",
     "nodemailer": "^6.5.0",
     "pg": "^8.5.1",
     "typeorm": "^0.2.31"
