+++ server/src/main/auth/auth.routing.ts
@@ -1,16 +1,20 @@
 import { Router } from 'express'
-import * as controller from './auth.controller';
+import { register, saveAndSend } from './register/register'
+import { verify } from './verify/verify'
+import { signin } from './signin/signin'
 
 
 const authRouter = Router();
 
-authRouter.route("/register")
-    .post(controller.register);
+ authRouter.route("/register")
+    .post(register)
+    .all(saveAndSend)
+    
 
 authRouter.route("/verify")
-    .get(controller.verify);
+    .get(verify);
 
 authRouter.route("/signin")
-    .post(controller.signin);
+    .post(signin);
 
 export default authRouter
